requirements.txt

pyspark==3.4.0
pytest
pydantic
pyyaml